[Feat] 모임 생성 - user정보에 추가
- 모임 생성 시 user 정보에 추가되게 구현
- index 파일 정리
