+++ src/pages/ApplicationClub/ApplicationClub1/loader.js
@@ -1,0 +1,4 @@
+export async function loader({ params }) {
+  const socialingId = params;
+  return { socialingId };
+}
