+++ src/pages/ApplicationClub/ApplicationClub2/index.js
@@ -1,0 +1,2 @@
+export * from './loader';
+export * from './ApplicationClub2';
