+++ src/pages/Chatting/ChatRoom/index.js
@@ -1,0 +1,1 @@
+export * from './ChatRoom';
