+++ src/pages/CreateBookReview/index.js
@@ -1,0 +1,1 @@
+export * from './CreateBookReview';
