+++ src/pages/CreateClub/CreateClub1/index.js
@@ -1,0 +1,1 @@
+export * from './CreateClub1';
