+++ src/pages/CreateClub/CreateClub2/index.js
@@ -1,0 +1,2 @@
+export * from './loader';
+export * from './CreateClub2';
