+++ src/pages/CreateClub/CreateClub3/index.js
@@ -1,0 +1,1 @@
+export * from './CreateClub3';
