+++ src/pages/CreateClub/CreateClub4/index.js
@@ -1,2 +1,1 @@
-export * from './loader';
 export * from './CreateClub4';
