+++ src/pages/DetailBookReview/DetailBookReview.jsx
@@ -1,11 +1,10 @@
-import pb from '@/api/pocketbase';
+import { NomalTitle, Svg } from '@/components/Atoms';
+import { Avatar, GNB } from '@/components/Molecules';
+import { getDocumentTitle } from '@/utils';
 import { Helmet } from 'react-helmet-async';
 import { useLoaderData } from 'react-router-dom';
-import { NomalTitle, Svg } from '@/components/Atoms';
-import { GNB, Avatar } from '@/components/Molecules';
-import { getPbImgs, getDocumentTitle } from '@/utils';
 
-function DetailBookReview() {
+export function DetailBookReview() {
   const {
     reviewInfo: {
       bookTitle,
@@ -30,7 +29,7 @@ function DetailBookReview() {
           <img className="object-cover" src={img} alt={title} />
         </figure>
         <Avatar nickName={nickname} text={title}></Avatar>
-        <section className="bg-bjgray-50 flex h-full flex-col gap-4 px-4 pt-10">
+        <section className="flex h-full flex-col gap-4 bg-bjgray-50 px-4 pt-10">
           <div className="flex justify-center gap-2 pt-[63px] text-b-3-light text-bjgray-500">
             <span className="flex items-center">
               <Svg color="#9e9e9e" size={12} id="book" />
@@ -45,15 +44,4 @@ function DetailBookReview() {
       </div>
     </>
   );
-}
-
-export default DetailBookReview;
-
-export async function loader({ params }) {
-  const { bookreviewId } = params;
-  const reviewInfo = await pb
-    .collection('bookReview')
-    .getOne(bookreviewId, { expand: 'writer, book_ISBN' });
-  reviewInfo.img = getPbImgs(reviewInfo);
-  return { reviewInfo };
 }
