+++ src/pages/DetailClub/index.js
@@ -1,0 +1,2 @@
+export * from './loader';
+export * from './DetailClub';
