+++ src/pages/Filter/index.js
@@ -1,0 +1,2 @@
+export * from './Filter';
+export * from './loader';
