+++ src/pages/Filter/loader.js
@@ -1,0 +1,6 @@
+import pb from '@/api/pocketbase';
+
+export async function loader() {
+  const filters = await pb.collection('genres').getFullList();
+  return filters;
+}
