+++ src/pages/MainBookReview/index.js
@@ -1,0 +1,2 @@
+export * from './loader';
+export * from './MainBookReview';
