+++ src/pages/MainClub/MainClub.jsx
@@ -1,42 +1,8 @@
-import pb from '@/api/pocketbase';
 import { Badge, MainButton, Svg } from '@/components/Atoms';
 import { GNB, MainKindToggle } from '@/components/Molecules';
-import { calcDay, getDocumentTitle, getPbImgs } from '@/utils';
+import { calcDay, getDocumentTitle } from '@/utils';
 import { Helmet } from 'react-helmet-async';
 import { Link, useLoaderData } from 'react-router-dom';
-
-export async function loader({ request }) {
-  const url = new URL(request.url);
-  const filters = url.searchParams.get('filters');
-
-  let filterQuery = null;
-
-  if (filters) {
-    if (filters.includes(',')) {
-      filterQuery = filters
-        .split(',')
-        .map((filter) => `genre.title~"${filter}"`)
-        .join('||');
-    } else {
-      filterQuery = `genre.title~"${filters}"`;
-    }
-  }
-
-  const clubs = await pb.collection('socialing').getFullList({
-    fields:
-      'id,title,dateTime,isOffline,collectionId,location,limitPerson,confirmUser,img,expand.genre.title',
-    expand: 'genre',
-    filter: filterQuery,
-    sort: '-created',
-  });
-
-  const clubItems = clubs.map((club) => {
-    const photoURL = getPbImgs(club);
-    club.photo = photoURL;
-    return club;
-  });
-  return clubItems;
-}
 
 function ClubCard() {
   const getClubList = useLoaderData();
@@ -89,7 +55,7 @@ function ClubCard() {
   );
 }
 
-function MainClub() {
+export function MainClub() {
   return (
     <>
       <Helmet>
@@ -128,5 +94,3 @@ function MainClub() {
     </>
   );
 }
-
-export default MainClub;
