+++ src/pages/MainClub/index.js
@@ -1,0 +1,2 @@
+export * from './loader';
+export * from './MainClub';
