+++ src/pages/MyClubList/index.js
@@ -1,0 +1,1 @@
+export * from './MyClubList';
