+++ src/pages/SignUp/index.js
@@ -1,0 +1,3 @@
+export * from './BasicInfo';
+export * from './DetailInfo';
+export * from './SignUp';
