+++ src/pages/Welcome/index.js
@@ -1,0 +1,1 @@
+export * from './Welcome';
