--- src/pages/index.js
@@ -1,22 +1,0 @@
-export { default as ApplicationClub1 } from '@/pages/ApplicationClub/ApplicationClub1/ApplicationClub1';
-export { default as ApplicationClub2 } from '@/pages/ApplicationClub/ApplicationClub2/ApplicationClub2';
-export { default as DetailBookReview } from '@/pages/DetailBookReview/DetailBookReview';
-export { default as CreateBookReview } from '@/pages/CreateBookReview/CreateBookReview';
-export { default as ChatRoomList } from '@/pages/Chatting/ChatRoomList/ChatRoomList';
-export { default as CreateClub1 } from '@/pages/CreateClub/CreateClub1/CreateClub1';
-export { default as CreateClub2 } from '@/pages/CreateClub/CreateClub2/CreateClub2';
-export { default as CreateClub3 } from '@/pages/CreateClub/CreateClub3/CreateClub3';
-// export { default as CreateClub4 } from '@/pages/CreateClub/CreateClub4/CreateClub4';
-export { default as MainBookReview } from '@/pages/MainBookReview/MainBookReview';
-// export { default as ManagementClub } from '@/pages/ManagementClub/ManagementClub';
-export { default as ChatRoom } from '@/pages/Chatting/ChatRoom/ChatRoom';
-export { default as MyClubList } from '@/pages/MyClubList/MyClubList';
-export { default as DetailClub } from '@/pages/DetailClub/DetailClub';
-export { default as Error404 } from '@/pages/Error404/Erro404';
-export { default as MainClub } from '@/pages/MainClub/MainClub';
-export { default as Welcome } from '@/pages/Welcome/Welcome';
-export { default as MyPage } from '@/pages/MyPage/MyPage';
-export { default as Search } from '@/pages/Search/Search';
-export { default as SignUp } from '@/pages/SignUp/SignUp';
-export { default as Intro } from '@/pages/Intro/Intro';
-export { default as Login } from '@/pages/Login/Login';
